Update main.js
Reformatted for clarity. Added new logic to editWidget process that displays a completion message which fades after a couple seconds. Currently it goes where the mouse was clicked, but needs to be changed to go where the point on the map is added ( or closer to it). Added tooltip to editWidget by making use of esri classes