Update main.js
Added visited variable that get's upated based on whether or not a place is visited. This is used in subsquent functions to updated the HTML on the map letting the user know whether a place has been visited or not. Got rid of the breakpoint stuff in addRankStars() and added the ability to add filled out of not filled stars.  Added creation of new elements to support the new visited variable.